requirements.txt
PyQt5==5.15.11
mysql-connector-python==9.2.0
python-dotenv==1.0.1
matplotlib==3.10.0
seaborn==0.13.2
numpy==2.2.3
pandas==2.2.3